requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
cachetools==4.2.4
celery==5.3.6
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.33.0
colorama==0.4.6
Django==4.2.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
django-session-timeout==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
factory-boy==3.3.0
Faker==20.1.0
google-api-core==2.10.2
google-api-python-client==2.97.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.60.0
httplib2==0.22.0
idna==3.4
kombu==5.3.4
Pillow==10.0.0
prompt-toolkit==3.0.43
protobuf==4.24.1
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.1.0
wcwidth==0.2.12
